requirements.txt
# AI SDK
google-genai==1.3.0

# RAG Infrastructure
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
pypdf==4.1.0

# Local Vector Storage
chromadb==0.5.3
sentence-transformers==2.6.1
huggingface-hub==0.22.2

# Frontend
streamlit==1.32.0